Implement authorization using Gates

+++ app/Http/Controllers/QuestionController.php
@@ -67,7 +67,10 @@ class QuestionController extends Controller
      */
     public function edit(Question $question)
     {
-
+        if (\Gate::denies('update-question', $question)) {
+            abort(403, 'Access denied!');
+        }
+        
         return view("question.edit", compact('question'));
     }
 
@@ -80,6 +83,10 @@ class QuestionController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if (\Gate::denies('update-question', $question)) {
+            abort(403, 'Access denied!');
+        }
+
         $question->update($request->only('title', 'body'));
 
         return redirect('/question')->with('success', "Your question has been updated.");
@@ -93,6 +100,10 @@ class QuestionController extends Controller
      */
     public function destroy(Question $question)
     {
+        if (\Gate::denies('delete-question', $question)) {
+            abort(403, 'Access denied!');
+        }
+
         $question->delete();
 
         return redirect('/question')->with('success', "Your question has been deleted.");
